Fix up availability behavior
Make the available field reflect the total number the current user could order, not factoring in what they currently have ordered. This makes the model much simpler and easier to think about.

Also, if the users gets a quantityNotAvailable error, include the quantity that is available in the error response.