Fix appending order rows
Apparently the sheets append() API behaves very unexpectedly (to me at least) if you give it a range that consists of a single hidden cell. Even if that cell is populated, it seems to just not look for an enclosing table and write starting at the hidden cell, overwriting it and other adjacent cells. We were using a cell in the first column of the orders sheet for our range, and since the script in the spreadsheet hides columns that don't have any availability, if the first product didn't have any availability, we'd hit this issue and not be able to append order rows correctly.

So, instead specify the first column of the first row where user orders get filled in -- it will either be blank if nobody has ordered, or will be be part of the "table" if people have, and will never be hidden.

So, instead pass in a range that extends across many columns to be sure that we catch at least one that isn't hidden.

foo