Remove mutex
The mutex is causing more trouble than it was worth.

Our google sheets API calls are rate-limited to about 1 per second (separately for writes and for reads), which generally shouldn't be a problem, but the mutex adds two writes anytime we're setting order quantities, and if there's trouble locking it, it will add a lot more calls as it retries. It's not too hard to exceed the quota and get rate-limited, and if that happens while when trying to clear the mutex, then the mutex can "leak" and get stuck locked, which is a very bad failure mode.

It's potentially possible to fix this with some more careful logic around the mutex, like perhaps doing a read before the append to try to avoid doing writes that will result in not locking the mutex, or using the timestamps written to the mutex table to detect and recover from this. But the mutex is only used to ensure the available quantites aren't exceeded by two concurrent orders neither of whom exceed the max, but whose sum does, and this can be dealt with by:

1. Hoping it doesn't happen very often (which it likely won't since the timing window is fairly narrow), and when it does rely on the farmers handling it "out-of-band" (contact the user and apologize or give them a substitute or something)
2. After setting the order, reading back the ordered quantities, and if they exceed the available quantities, do another write to revert the order
3. Look into data validations in Google Sheets to see if Google Sheets can do the work for us

Regardless, this reduces the failure mode from all users being unable to place oders to the order quantities exceeding the availability, but (we expect) only *very* rarely. We're (de-factor) going with option (1) now, but if that proves to be problematic, we can look into (2) or (3).